requirements.txt
requests
beautifulsoup4
pandas
elasticsearch
python-dotenv
fastapi
uvicorn
